Move editor element components into folder
