+++ components/editor/Editor.tsx
@@ -37,11 +37,10 @@ import { useStore } from 'lib/store';
 import { ElementType, Mark } from 'types/slate';
 import HoveringToolbar from './HoveringToolbar';
 import AddLinkPopover from './AddLinkPopover';
-import EditorElement, {
-  withOptionsMenu,
-  withVerticalSpacing,
-} from './EditorElement';
-import EditorLeaf from './EditorLeaf';
+import EditorElement from './elements/EditorElement';
+import { withVerticalSpacing } from './elements/withVerticalSpacing';
+import { withOptionsMenu } from './elements/withOptionsMenu';
+import EditorLeaf, { EditorLeafProps } from './elements/EditorLeaf';
 import LinkAutocompletePopover from './LinkAutocompletePopover';
 
 export type AddLinkPopoverState = {
@@ -82,7 +81,10 @@ export default function Editor(props: Props) {
     );
     return EditorElementWithOptionsMenu;
   }, []);
-  const renderLeaf = useCallback((props) => <EditorLeaf {...props} />, []);
+  const renderLeaf = useCallback(
+    (props: EditorLeafProps) => <EditorLeaf {...props} />,
+    []
+  );
 
   const [addLinkPopoverState, setAddLinkPopoverState] =
     useState<AddLinkPopoverState>({
